kernel: fix overflow problem due to wrong type of expires field

+++ kernel/rlite-kernel.h
@@ -438,7 +438,7 @@ struct flow_entry {
 
     struct rl_flow_stats stats;
     struct list_head    node_rm; /* for flows_removeq */
-    unsigned            expires; /* absolute time in jiffies */
+    unsigned long       expires; /* absolute time in jiffies */
     unsigned int        refcnt;
 #define RL_FLOW_NEVER_BOUND     (1 << 0) /* flow was never bound with ioctl */
 #define RL_FLOW_PENDING         (1 << 1) /* flow allocation is pending */
